% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ulits.R
\name{canonicalForm}
\alias{canonicalForm}
\title{Canonical Form}
\usage{
canonicalForm(x, y, V)
}
\arguments{
\item{x}{vector}

\item{y}{observations}

\item{V}{a symmetrix matrix whoes eigenvalues all lie in [0, 1]}
}
\description{
Canonical Form
}

