% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\docType{class}
\name{Cube-class}
\alias{Cube-class}
\title{Class \code{"Cube"}}
\description{
Class \code{"Cube"}
}
\section{Slots}{

\describe{
\item{\code{data}}{(array) The data that are represented as hypercube.}

\item{\code{structure}}{(list) The structure of the dimensions of the hypercube.}

\item{\code{view}}{(list) Information about how to build a view for the hypercube. This information is stored in a list of \code{\link{Dimension-class}} objects.}
}}

\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("Cube", ...)}. This S4 class describes \code{Cube} objects.
}

\examples{

# show Cube definition
showClass("Cube")

}
\seealso{
\code{\link{generateCube}}
}
\author{
Michael Scholz \email{michael.scholz@th-deg.de}
}
\keyword{classes}
