% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension.R
\docType{class}
\name{Dimension-class}
\alias{Dimension-class}
\title{Class \code{"Cube"}}
\description{
Class \code{"Cube"}
}
\section{Slots}{

\describe{
\item{\code{name}}{(character) The name of the dimension.}

\item{\code{values}}{(vector) A vector of selected values for this dimension.}

\item{\code{aggregation}}{(vector) A vector of aggregation functions that will be applied to this dimension.}
}}

\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("Dimension", ...)}. This S4 class describes \code{Dimension} objects.
}

\examples{

# show Dimension definition
showClass("Dimension")


}
\author{
Michael Scholz \email{michael.scholz@th-deg.de}
}
\keyword{classes}
