% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ulits.R
\name{estRiskCanonical}
\alias{estRiskCanonical}
\title{Estimate Risk Canonical}
\usage{
estRiskCanonical(canonicalform, estsig)
}
\arguments{
\item{canonicalform}{canonical form}

\item{estsig}{estimated variance}
}
\description{
Estimate Risk Canonical
}

