% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hypercubeEst.R
\name{fitted.hypercube}
\alias{fitted.hypercube}
\title{Fitted values of the class "\code{hypecube}".}
\usage{
\method{fitted}{hypercube}(object, ...)
}
\description{
Fitted values of the class "\code{hypecube}".
}

