\name{HyperCube-package}
\alias{HyperCube-package}
\alias{HyperCube}
\docType{package}
\title{
Hypercube estimator
}
\description{
More about what it does (maybe more than one line)
}
\details{
\tabular{ll}{
Package: \tab HyperCube\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-05-08\cr
License: \tab BSD\cr
}
}
\author{
Chi Po Choi
Amy Kim

Maintainer: Chi Po Choi <cpchoi@ucdavis.edu>, Amy Kim <atykimi@ucdavis.edu>, 
}
\references{ Rudy 
}
\keyword{ package }
\seealso{
}
\examples{
}
