% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hypercubeEst.R
\name{hypercubeOp}
\alias{hypercubeOp}
\title{Hypercube Operator}
\usage{
hypercubeOp(X, V)
}
\arguments{
\item{X}{design matrix}

\item{V}{sysmmetric matrix whose eigenvalues all lie in [0,1]}
}
\description{
Hypercube Operator
}
\references{
Beran, Rudolf. "Hypercube estimators: Penalized least squares, submodel selection, and numerical stability." Computational Statistics & Data Analysis 71 (2014): 654-666.
}

