% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hypercubeEst.R
\name{hypercubeOptimization}
\alias{hypercubeOptimization}
\title{Hypercube Optimization}
\usage{
hypercubeOptimization(formula, data, sigma = NULL)
}
\arguments{
\item{formula}{formula}

\item{data}{data}

\item{sigma}{estimated variance}
}
\description{
give the component which minimizing the risk
}
\references{
Beran, Rudolf. "Hypercube estimators: Penalized least squares, submodel selection, and numerical stability." Computational Statistics & Data Analysis 71 (2014): 654-666.
}

