% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ulits.R
\name{modelMatrix}
\alias{modelMatrix}
\title{Generate model matrix}
\usage{
modelMatrix(formula, data)
}
\arguments{
\item{formula}{formula}

\item{data}{data}
}
\description{
generate model matrix given formula and data
}

