% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monkey}
\alias{monkey}
\title{Response of 5 different monkey-pairs}
\format{A data frame with 25 rows and 4 variables:
\describe{
  \item{Cond.n}{the condition}
  \item{Monkeys}{the monkey pair}
  \item{Period}{the monkey pair}
  \item{Response}{responses to a certain stimulus}
}}
\source{
p. 189 of Scheffes text and reformatted in monkey.RData
}
\usage{
monkey
}
\description{
A dataset containing reports responses to a certain stimulus
that were measured for 5 different monkey-pairs (the subjects)
in 5 different periods under 5 different conditions
}
\keyword{datasets}

