% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{motor}
\alias{motor}
\title{Accelations over time}
\format{A data frame with 133 rows and 2 variables:
\describe{
  \item{t}{possible observation times}
  \item{accel}{acceleration against time}
}}
\source{
adapted from Silverman (1985)
}
\usage{
motor
}
\description{
A dataset containing 133 observation of motorsycle
acceleration against time in a simulated motorcycle accident.
The p = 277 possible observation times constitute the vector t = (1, 2, ... , 277)
Accelerations were observed at only q < p of these equally spaced time,
sometimes with replication.
}
\keyword{datasets}

