% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hypercubeEst.R
\name{predict.hypercube}
\alias{predict.hypercube}
\title{Predict method for Hypercube Estimator Fits}
\usage{
\method{predict}{hypercube}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{an object of the class "\code{hypercube}"}

\item{newdata}{new data}

\item{...}{other arguments}
}
\description{
Predict method for Hypercube Estimator Fits
}

