% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hypercubeEst.R
\name{print.hypercube}
\alias{print.hypercube}
\title{Print methods for hypercube}
\usage{
\method{print}{hypercube}(x, ...)
}
\arguments{
\item{x}{an object of class "\code{hypercube}"}

\item{...}{other arguments}
}
\description{
Print methods for hypercube
}

