% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{print.Importances}
\alias{print.Importances}
\title{Prints an Importances object.}
\usage{
\method{print}{Importances}(x, ...)
}
\arguments{
\item{x}{The \code{Importances} object that will be printed.}

\item{...}{Ignored parameters.}
}
\value{
Sparsity value
}
\description{
Prints an \code{Importances} object.
}
\examples{

data("sales")
cube = generateCube(sales, columns = list(time = c("month", "year"),
      location = c("state"), product = "product"), valueColumn = "amount")
importances = importance(cube)
print(importances)

}
\seealso{
\code{\link{importance}}
}
\author{
Michael Scholz \email{michael.scholz@uni-passau.de}
}
