% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hypercubeEst.R
\name{print.summary.hypercube}
\alias{print.summary.hypercube}
\title{Print methods for summary.hypercube}
\usage{
\method{print}{summary.hypercube}(x, ...)
}
\arguments{
\item{x}{an object of class "\code{summary.hypercube}"}

\item{...}{other arguments}
}
\description{
Print methods for summary.hypercube
}

