% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/projection.R
\name{projectCreate}
\alias{projectCreate}
\alias{projectFun}
\alias{projectSub}
\alias{projectWeight}
\title{Methods for object "\code{projection.hypercube}"}
\usage{
projectCreate(formula, data)

projectSub(proj, component)

projectWeight(proj, component = NULL, weights = NULL)

projectFun(proj, component = NULL)
}
\arguments{
\item{formula}{formula}

\item{data}{data}

\item{proj}{proj}

\item{component}{component}

\item{weights}{weights}
}
\description{
Methods for object "\code{projection.hypercube}"
}
\section{Functions}{
\itemize{
\item \code{projectSub}: Subset of object "\code{projection.hypercube}"

\item \code{projectWeight}: Weighted sum of matrix in object "\code{projection.hypercube}"

\item \code{projectFun}: Functions of object "\code{projection.hypercube}"
}}
\examples{
proj <- projectCreate( ~ mother:infant -1, data = litter)
}

