% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sales}
\alias{sales}
\title{Sales of books}
\format{A data fram with 2500 rows and 7 variables:
\describe{
  \item{month}{month as number}
  \item{year}{year as number}
  \item{state}{abbreviation of the state as character}
  \item{country}{country as character}
  \item{product}{name of the product as character}
  \item{unit}{number of sold products}
  \item{amount}{amount of sales}
}}
\source{
Synthetic dataset
}
\usage{
sales
}
\description{
A dataset containing 2,500 sales of 4 books in different states and countries.
}
\keyword{datasets}
