% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension.R
\docType{methods}
\name{show,Dimension-method}
\alias{show,Dimension-method}
\title{Shows a \code{Dimension} object}
\usage{
\S4method{show}{Dimension}(object)
}
\arguments{
\item{object}{The \code{Dimension} object}
}
\description{
Shows a \code{Dimension} object
}
\examples{

data("sales")
cube = generateCube(sales, columns = list(time = c("month", "year"),
      location = c("state"), product = "product"), valueColumn = "amount")
cube@view[[1]]

}
\seealso{
\code{\link[=Cube-class]{Cube}}
}
\author{
Michael Scholz \email{michael.scholz@uni-passau.de}
}
