% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,Cube-method}
\title{Shows a summary for the given cube}
\usage{
summary(x)
}
\arguments{
\item{x}{The \code{Cube} object for which the summary is shown.}
}
\description{
Shows the dimensions and the number of levels per dimension of the given cube. All added selections and aggregations will be
regarded.
}
\examples{

data("sales")
cube = generateCube(sales, columns = list(time = c("month", "year"),
      location = c("state"), product = "product"), valueColumn = "amount")
summary(cube)

}
\seealso{
\code{\link[=Cube-class]{Cube}}
}
\author{
Michael Scholz \email{michael.scholz@th-deg.de}
}
\keyword{methods}
