% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vineyard}
\alias{vineyard}
\title{Vineyard}
\format{A data frame with 52 rows and 4 variables:
\describe{
  \item{row}{the vineyard row number}
  \item{year1}{reporting the harvest yield in first year}
  \item{year2}{reporting the harvest yield in second year}
  \item{year3}{reporting the harvest yield in third year}
}}
\source{
vineyard
}
\usage{
vineyard
}
\description{
A dataset containing records the grape yield harvested in each row of a vinyard
in three succes years
}
\keyword{datasets}

