% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hypercubeEst.R
\name{hypercube}
\alias{hypercube}
\alias{hypercube.default}
\alias{hypercube.formula}
\title{Hypercube generic}
\usage{
hypercube(...)

\method{hypercube}{default}(X, y, V, ...)

\method{hypercube}{formula}(formula, data, V, ...)
}
\arguments{
\item{...}{Object to be hypercube}

\item{X}{design matrix}

\item{y}{observation}

\item{V}{sysmmetric matrix whose eigenvalues all lie in [0,1]}

\item{formula}{formula to get estimate}

\item{data}{data you want to analysis}

\item{...}{other optional arguments}
}
\description{
Hypercube generic

Hypercube Estimate
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: 

\item \code{formula}: 
}}
\references{
Beran, Rudolf. "Hypercube estimators: Penalized least squares, submodel selection, and numerical stability." Computational Statistics & Data Analysis 71 (2014): 654-666.
}

