% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HyperCube.R
\docType{package}
\name{HyperCube}
\alias{HyperCube}
\alias{HyperCube-package}
\title{Hypercube estimator}
\description{
Linear model with hypercube estimator, penalized least square and more.
}
\details{
\tabular{ll}{
Package: \tab HyperCube\cr
Type: \tab Package\cr
Version: \tab 0.1-1\cr
Date: \tab 2015-05-08\cr
License: \tab GPL\cr
}
}
\author{
Chi Po Choi \email{cpchoi@ucdavis.edu},
Amy Kim \email{atykimi@ucdavis.edu}

Maintainer: Chi Po Choi \email{cpchoi@ucdavis.edu}, Amy Kim \email{atykimi@ucdavis.edu}
}

