% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hypercubeEst.R
\name{hypercube}
\alias{hypercube}
\alias{hypercube.default}
\alias{hypercube.formula}
\title{Hypercube Estimator Fits}
\usage{
hypercube(...)

\method{hypercube}{default}(X, y, V, ...)

\method{hypercube}{formula}(formula, data, V, ...)
}
\arguments{
\item{...}{other optional arguments}

\item{X}{design matrix}

\item{y}{observation}

\item{V}{sysmmetric matrix whose eigenvalues all lie in [0,1]}

\item{formula}{formula to get estimate}

\item{data}{data you want to analysis}
}
\value{
A object of the class "\code{hypercube}" containing the following:
\describe{
  \item{coefficients}{Estimated coefficents \eqn{\hat{\beta}} under hypercube estimator}
  \item{fitted.values}{Fitted values (\eqn{\hat{\eta}})  under hypercube estimator}
  \item{residuals}{residuals, \eqn{y - \hat{\eta}}}
  \item{estsigma2}{Estimated variance of \eqn{\epsilon}.
   If the rank of the design matrix X is greater than the number of observation,
   the variance is estimated using least square regression.
   If the rank is equal to the number of observation,
   and if the model has more than one factor,
   the variance is estimated using additive submodel fit.
   If neither of the above cases,
   users need to provide their own estimated variance.}
  \item{estrisk}{Esitmated risk, which involved the estimated variance.}
  \item{V}{Symmtric matrix with all eigenvalues lie in [0,1].
  The matrix V used the hypercube estimator. }
  \item{modelframe}{The modelframe constructed from the argument formula and data.}
}
}
\description{
The user interface for fitting linear model with Hypercbue Estimator.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default method for \code{hypercube}

\item \code{formula}: Formula method for \code{hypercube}
}}
\note{
This package is still under development. For data with degree of freedom equal to zero, it may not be handled well. It is due the fact that it is not obvious to estimate the variance of the random error. Users need to provide estimated variance of random error in such cases.
}
\examples{
## Example 1 in Beran (2014)
## Fitting Canadian earning data with Hypercube Estimator

# The number of age, p, in Example 1 in Beran (2014).
p <- length(unique(canadian.earnings[,1]))
# D_5 as in equation (3.10) in Beran (2014)
D <- diffMatrix(p, 5)
# The parametor nu in equation (3.11) in Beran (2014)
nu <- 100
# The matrix W in equation (3.11) in Beran (2014)
W <- nu * t(D) \%*\% D
# Convert W to V, as described in (1.6) in Beran (2014)
V <- plsW2V(W)
# The variable age should be considered as a factor
canadian.earnings[,"age"] <- factor(canadian.earnings[,"age"])
# Hyperpercube Estimator Fit
hcmod <- hypercube( log.income ~ age -1, data=canadian.earnings, V)

# Plot of data
plot(as.numeric(as.character(canadian.earnings$age)),
     canadian.earnings$log.income,
     xlab = "age", ylab = "log(income)")
# Plot of fitted line
lines(levels(canadian.earnings$age), hcmod$coefficients)


## Example 2 in Beran (2014)
## Fitting rat litter data

# Projection matrices as decribed in equation (5.4) in Beran (2014)
litter.proj <- projectCreate( ~ mother:infant -1, data = litter)
# If only additive effect is consider,
# take V = P1 + P2 + P3 (notation in equation (5.4) in Beran (2014))
component <- cbind(c(0,0), c(1,0), c(0,1))
V <- projectWeight(litter.proj, component = component)
# Hypercube Estimator Fit
hcmod <- hypercube( weight ~ mother:infant -1, data = litter, V)

# Estimated Risk
summary(hcmod)

## Hypercube Estimator with optimal risk
##
hcmodopt <- hypercubeOptimization( weight ~ mother:infant -1,
                                   data = litter
                                   )
# The optimal projection coefficient which minimizes the risk.
hcmodopt$projcoef

# The minimum risk
hcmodopt$estrisk

# The Hypercube Estimator fit with the V of the optimal projection.
summary(hcmodopt$est)
}
\references{
Beran, Rudolf. "Hypercube estimators: Penalized least squares, submodel selection, and numerical stability." Computational Statistics & Data Analysis 71 (2014): 654-666.
}
\seealso{
\code{\link{hypercubeOptimization}}
}

