% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monkey}
\alias{monkey}
\title{Response of 5 different monkey-pairs}
\format{A data frame with 25 rows and 4 variables:
\describe{
  \item{cond}{the condition}
  \item{monkeys}{the monkey pair}
  \item{period}{the monkey pair}
  \item{response}{responses to a certain stimulus}
}}
\source{
Data from Query no. 113, edited by G.W. Sender, Biometrics, Vol. 11, 1955, p.112
}
\usage{
monkey
}
\description{
A dataset containing reports responses to a certain stimulus
that were measured for 5 different monkey-pairs (the subjects)
in 5 different periods under 5 different conditions
}
\keyword{datasets}

